Extract ListenerTracker, so that it can be shared
