+++ pkg/tracker/porttracker.go
@@ -1,5 +1,5 @@
 /*
-Copyright © 2022 SUSE LLC
+Copyright © 2023 SUSE LLC
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -16,18 +16,9 @@ limitations under the License.
 package tracker
 
 import (
-	"context"
-	"errors"
-	"net"
-	"strconv"
-	"sync"
-	"syscall"
-
-	"github.com/Masterminds/log-go"
 	"github.com/docker/go-connections/nat"
 	"github.com/rancher-sandbox/rancher-desktop-agent/pkg/forwarder"
 	"github.com/rancher-sandbox/rancher-desktop-agent/pkg/types"
-	"golang.org/x/sys/unix"
 )
 
 type Tracker interface {
@@ -35,8 +26,7 @@ type Tracker interface {
 	Add(containerID string, portMapping nat.PortMap) error
 	Remove(containerID string) error
 	RemoveAll()
-	AddListener(ctx context.Context, ip net.IP, port int) error
-	RemoveListener(ctx context.Context, ip net.IP, port int) error
+	NetTracker
 }
 
 // PortTracker keeps track of port mappings and forwards
@@ -46,9 +36,7 @@ type PortTracker struct {
 	portStorage      *portStorage
 	vtunnelForwarder *forwarder.VtunnelForwarder
 	wslAddrs         []types.ConnectAddrs
-	// outstanding listeners; the key is generated via ipPortToAddr.
-	listeners map[string]net.Listener
-	mutex     sync.Mutex
+	*ListenerTracker
 }
 
 // NewPortTracker creates a new Port Tracker.
@@ -57,7 +45,7 @@ func NewPortTracker(forwarder *forwarder.VtunnelForwarder, wslAddrs []types.Conn
 		portStorage:      newPortStorage(),
 		vtunnelForwarder: forwarder,
 		wslAddrs:         wslAddrs,
-		listeners:        make(map[string]net.Listener),
+		ListenerTracker:  NewListenerTracker(),
 	}
 }
 
@@ -110,124 +98,4 @@ func (p *PortTracker) Remove(containerID string) error {
 // RemoveAll removes all the port bindings from the tracker.
 func (p *PortTracker) RemoveAll() {
 	p.portStorage.removeAll()
-}
-
-// AddListener adds an IP / port combination into the listener tracker.
-// If this combination is already being tracked, this is a no-op.
-func (p *PortTracker) AddListener(ctx context.Context, ip net.IP, port int) error {
-	addr := ipPortToAddr(ip, port)
-
-	listener, err := listen(ctx, addr)
-	if err != nil {
-		return err
-	}
-
-	p.mutex.Lock()
-	p.listeners[addr] = listener
-	p.mutex.Unlock()
-
-	return nil
-}
-
-// RemoveListener removes an IP / port combination from the listener tracker.  If this
-// combination was not being tracked, this is a no-op.
-func (p *PortTracker) RemoveListener(_ context.Context, ip net.IP, port int) error {
-	addr := ipPortToAddr(ip, port)
-
-	p.mutex.Lock()
-	defer p.mutex.Unlock()
-
-	if listener, ok := p.listeners[addr]; ok {
-		if err := listener.Close(); err != nil {
-			return err
-		}
-
-		delete(p.listeners, addr)
-	}
-
-	return nil
-}
-
-func ipPortToAddr(ip net.IP, port int) string {
-	return net.JoinHostPort(ip.String(), strconv.Itoa(port))
-}
-
-// Listen on the given address and port.  The returned listener never handles
-// any traffic (immediately closing any incoming connection), and tries to
-// shutdown quickly when no longer needed.
-func listen(ctx context.Context, addr string) (net.Listener, error) {
-	config := &net.ListenConfig{
-		Control: func(network, address string, c syscall.RawConn) error {
-			//nolint:varnamelen // `fd` is the typical name for file descriptor
-			err := c.Control(func(fd uintptr) {
-				// We should never get any traffic, and should
-				// never wait on close; so set linger timeout to
-				// 0.  This prevents normal socket close, but
-				// that's okay as we don't handle any traffic.
-				err := unix.SetsockoptLinger(int(fd), unix.SOL_SOCKET, unix.SO_LINGER, &unix.Linger{
-					Onoff:  1,
-					Linger: 0,
-				})
-				if err != nil {
-					log.Errorw("failed to set SO_LINGER", log.Fields{
-						"error": err,
-						"addr":  addr,
-						"fd":    fd,
-					})
-				}
-				err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEADDR, 1)
-				if err != nil {
-					log.Errorw("failed to set SO_REUSEADDR", log.Fields{
-						"error": err,
-						"addr":  addr,
-						"fd":    fd,
-					})
-				}
-				err = unix.SetsockoptInt(int(fd), unix.SOL_SOCKET, unix.SO_REUSEPORT, 1)
-				if err != nil {
-					log.Errorw("failed to set SO_REUSEPORT", log.Fields{
-						"error": err,
-						"addr":  addr,
-						"fd":    fd,
-					})
-				}
-			})
-			if err != nil {
-				return err
-			}
-
-			return nil
-		},
-	}
-
-	listener, err := config.Listen(ctx, "tcp4", addr)
-	if err != nil {
-		return nil, err
-	}
-
-	go func() {
-		for {
-			conn, err := listener.Accept()
-			if err != nil {
-				if !errors.Is(err, net.ErrClosed) {
-					log.Errorw("failed to accept connection", log.Fields{
-						"error": err,
-						"addr":  addr,
-					})
-				}
-
-				return
-			}
-			// We don't handle any traffic; just unceremoniously
-			// close the connection and let the other side deal.
-			if err = conn.Close(); err != nil {
-				log.Errorw("failed to close connection", log.Fields{
-					"error": err,
-					"addr":  addr,
-				})
-			}
-		}
-	}()
-
-	return listener, nil
 }
